+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 11.12.4
+ - Changed the `manage_template` default value to `false` when data streams is enabled [#1111](https://github.com/logstash-plugins/logstash-output-elasticsearch/pull/1111)
+    -  Added the `manage_template => false` as a valid data stream option
+
 ## 11.12.3
 - Changed the log messages for data stream checks [#1109](https://github.com/logstash-plugins/logstash-output-elasticsearch/pull/1109)
     - Added more details about incompatible data streams supplied configurations